requirements.txt
Django==2.0.2
djangorestframework==3.9.0
mysqlclient==1.3.14
pandas==0.23.4
Pillow==5.2.0